Polishing

+++ lib/debugging/config.py
@@ -4,7 +4,6 @@ import os
 from lib.debugging.log import write_log
 from lib.debugging.subdirectory import Subdirectory
 
-# Absolute pathing
 script_path = os.path.abspath(__file__)
 config_file_path = os.path.join(script_path, '../../../config.ini')
 
@@ -13,16 +12,28 @@ config.read(config_file_path)
 
 
 def get_config(section: str, option: str) -> str:
-    try:
-        return config.get(section, option)
-    except (configparser.NoSectionError, configparser.NoOptionError) as err:
-        write_log(err, Subdirectory.CONFIG)
+    if os.path.exists(config_file_path):
+        write_log(f"get_config(): Configuration file {config_file_path} could not be found", Subdirectory.CONFIG)
+    else:
+        try:
+            return config.get(section, option)
+        except (configparser.NoSectionError, configparser.NoOptionError) as err:
+            write_log(f"get_config(): {err}", Subdirectory.CONFIG)
 
 
 def change_config(section: str, option: str, value: str) -> None:
-    config.set(section, option, value)
+    if os.path.exists(config_file_path):
+        write_log(f"change_config(): Configuration file {config_file_path} could not be found", Subdirectory.CONFIG)
+    else:
+        try:
+            config.set(section, option, value)
+        except (configparser.NoSectionError, configparser.NoOptionError) as err:
+            write_log(f"change_config(): {err}", Subdirectory.CONFIG)
 
 
 def write_config():
-    with open('config.ini', 'w') as config_file:
-        config.write(config_file)
+    if os.path.exists(config_file_path):
+        write_log(f"write_config(): Configuration file {config_file_path} could not be found", Subdirectory.CONFIG)
+    else:
+        with open('config.ini', 'w') as config_file:
+            config.write(config_file)
